+++ chat/serializers.py
@@ -18,7 +18,7 @@ class ChatSerializer(serializers.ModelSerializer):
 class PostMessageSerializer(serializers.ModelSerializer):
     class Meta:
         model = Message
-        fields = ("PGP_message", "order", "offset")
+        fields = ("PGP_message", "order_id", "offset")
         depth = 0
 
     offset = serializers.IntegerField(
